requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
asyncpg==0.29.0
sqlalchemy==2.0.31
alembic==1.13.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.3.4
Pillow==10.4.0
python-dotenv==1.0.1